# OpenApiConfig/openapi-spec.yaml
openapi: 3.0.0
paths:
...
  /grinders/{id}:
    get:
      operationId: GrindersController_get
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Grinder Successfully Read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadGrinderDto"
        "400":
          description: Bad request
        "500":
          description: Internal Server Error
      summary: Get a Grinder
      tags: &a4
        - Grinders
    delete:
      operationId: GrindersController_delete
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: grinder Successfully Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadGrinderDto"
        "400":
          description: Bad request
        "500":
          description: Internal Server Error
      summary: Delete One Grinder
      tags: *a4
    patch:
      operationId: GrindersController_edit
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EditGrinderDto"
      responses:
        "200":
          description: Grinder Successfully edited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadGrinderDto"
        "400":
          description: Bad request
        "500":
          description: Internal Server Error
      summary: Edit One Grinder
      tags: *a4
  /grinders/many/{id}:
    get:
      operationId: GrindersController_getAll
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Grinders Successfully Read
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReadGrinderDto"
        "400":
          description: Bad request
        "500":
          description: Internal Server Error
      summary: Get many Grinders
      tags: *a4
  /grinders:
    post:
      operationId: GrindersController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGrinderDto"
      responses:
        "200":
          description: Grinder Successfully Created
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReadGrinderDto"
        "400":
          description: Bad request
        "500":
          description: Internal Server Error
      summary: Create One Grinder
      tags: *a4
...
  /user/user:
    patch:
      operationId: UsersController_register
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EditUserDto"
      responses:
        "201":
          description: User successfully edited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponseDto"
        "400":
          description: Bad request
        "500":
          description: Internal Server Error
      summary: edit a user
      tags: &a6
        - user
    delete:
      operationId: UsersController_login
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteUserDto"
      responses:
        "200":
          description: Delete User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponseDto"
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      summary: Delete a user
      tags: *a6
info:
  title: SyncSipApi
  description: API For the SyncSip App
  version: "1.0"
  contact: {}
...
servers: []
components:
  schemas:
...
    UserResponseDto:
      type: object
      properties:
        id:
          type: string
          description: The user's id
          example: "234957689234"
        name:
          type: string
          description: The user's name
          example: John Doe
      required:
        - id
        - name
...
    ReadGrinderDto:
      type: object
      properties:
        id:
          type: string
          description: The grinder's id
          example: 123e4567-e89b-12d3-a456-426614174000
          format: uuid
        brandName:
          type: string
          description: The grinder's brand name
          example: Niche
        model:
          type: string
          description: The grinder's model
          example: Zero
        userId:
          type: string
          description: The user's id the grinder belongs to
          example: 123e4567-e89b-12d3-a456-426614174000
          format: uuid
      required:
        - id
        - brandName
        - model
        - userId
    CreateGrinderDto:
      type: object
      properties:
        brandName:
          type: string
          description: The grinder's brand name
          example: Niche
        model:
          type: string
          description: The grinder's model
          example: Zero
        userId:
          type: string
          description: The user's id the grinder belongs to
          example: 123e4567-e89b-12d3-a456-426614174000
          format: uuid
      required:
        - brandName
        - model
        - userId
    EditGrinderDto:
      type: object
      properties:
        brandName:
          type: string
          description: The grinder's brand name
          example: Niche
        model:
          type: string
          description: The grinder's model
          example: Zero
...
    EditUserDto:
      type: object
      properties:
        name:
          type: string
          description: The user's name
          example: John Doe
      required:
        - name
    DeleteUserDto:
      type: object
      properties:
        id:
          type: string
          description: The users id
          example: "76342976324893"
      required:
        - id